[SPARK-37147][SS] MetricsReporter producing NullPointerException when element 'triggerExecution' not present in Map[]
### What changes were proposed in this pull request?

Bug Fix.

The problematic code is in `MetricsReporter`:
`registerGauge("latency", _.durationMs.get("triggerExecution").longValue(), 0L)`
Instead of `.getOrDefault(...).longValue()` it uses `.get("triggerExecution").longValue()` which can return a null pointer exception if "triggerExecution" is not in the durationMs map.

Solution: use `.getOrDefault` when accessing a map.

### Why are the changes needed?

When `MetricsReporter.scala` registers a Gauge it occasionally returns a NPE. This breaks reporting custom metrics via Dropwizard and logs multiple times a stacktrace. It usually happens when using StructuredStreaming on a slow data source but I'm not able to reliably reproduce it every time.

### Does this PR introduce _any_ user-facing change?

Yes - fixes occasional failures when reporting metrics with Dropwizard

### How was this patch tested?

Added a unit-test.

Closes #34426 from gitplaneta/SPARK-37147.

Authored-by: Radek Busz <ext.radek.busz@ef.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>