[SPARK-37388][SQL] Fix NPE in WidthBucket in WholeStageCodegenExec
This PR fixes a `NullPointerException` in `WholeStageCodegenExec` caused by the expression `WidthBucket`. The cause of this NPE is that `WidthBucket` calls `WidthBucket.computeBucketNumber`, which can return `null`, but the generated code cannot deal with `null`s.

This fixes a `NullPointerException` in Spark SQL.

No

Added tests to `MathExpressionsSuite`. This suite already had tests for `WidthBucket` with interval inputs, but lacked tests with double inputs. I checked that the tests failed without the fix, and succeed with the fix.

Closes #34670 from tomvanbussel/SPARK-37388.

Authored-by: Tom van Bussel <tom.vanbussel@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>
(cherry picked from commit 77f397464823a547c27d98ed306703cb9c73cec3)
Signed-off-by: Wenchen Fan <wenchen@databricks.com>